Update on scenarios and expanded

+++ backend/src/main/java/com/ego/ethicai/dto/practice/DecisionTreeAlternativeDTO.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class DecisionTreeAlternativeDTO {
     private String text;
     private String tactic;
-    private Integer evs;
+    private Double evs;
 }
